+++ Documentation/3.0/Raven.Documentation.Pages/start/whats-new.markdown
@@ -21,6 +21,7 @@
 ### Client API
 
 - added `PreserveDocumentPropertiesNotFoundOnModel` convention. More [here](../client-api/configuration/conventions/request-handling#preservedocumentpropertiesnotfoundonmodel).
+- **highlights** can be accessed when performing **projection** or querying **map-reduce** index. More [here](../indexes/querying/highlights#highlights--projections).
 
 {PANEL/}